第一项目的第三阶段完成。

+++ Simple Chatty Bot/Problems/National assembly/src/Main.kt
@@ -4,6 +4,8 @@ import kotlin.math.roundToInt
 
 fun main(args: Array<String>) {
     val scanner = Scanner(System.`in`)
-    val population = scanner.nextDouble()
-    println(floor(Math.cbrt(population)).roundToInt())
+
+    val population = scanner.nextInt()
+
+    println(floor(Math.cbrt(population.toDouble())).roundToInt())
 }
